+++ src/Traits/HasLogins.php
@@ -15,9 +15,37 @@ trait HasLogins
     public function logins(): HasMany
     {
         return $this->hasMany(
-            config('model-login.model', Login::class),
+            config('model-login.model'),
             'user_id',
             'id'
         );
+    }
+
+    /**
+     * Successful Logins relation.
+     *
+     * @return HasMany
+     */
+    public function successful_logins(): HasMany
+    {
+        return $this->hasMany(
+            config('model-login.model'),
+            'user_id',
+            'id'
+        )->whereStatus(Login::STATUS_SUCCESSFUL);
+    }
+
+    /**
+     * Failed Logins relation.
+     *
+     * @return HasMany
+     */
+    public function failed_logins(): HasMany
+    {
+        return $this->hasMany(
+            config('model-login.model'),
+            'user_id',
+            'id'
+        )->whereStatus(Login::STATUS_FAILED);
     }
 }