+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.2 - 2021-02-15
+
+- add guard and remove identifier
+- ensure user has logins to perform log
+- fix typo for successful 
+- add successful/failed logins relation
+
 ## 1.0.1 - 2021-02-12
 
 - remove failed_logins relation